# Hw10 program optimization is completed