--- a/hw10_program_optimization/go.mod
+++ b/hw10_program_optimization/go.mod
@@ -4,7 +4,8 @@ go 1.16
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pquerna/ffjson v0.0.0-20190930134022-aa0246cd15f7
 	github.com/stretchr/testify v1.7.0
-	gopkg.in/yaml.v2 v2.4.0 // indirect
+	github.com/valyala/fastjson v1.6.3
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
 )